Remove Model.[] and Model.[]=
These two methods were simple aliases for @identity_map[] and []=.
It was bad practice to access models in this fashion, so this was
removed. Model.identity_map is still available to access models
directly, but Model.find() is recomended as it will fetch a remote model
if a local one is not available.